+++ example_runner.py
@@ -9,6 +9,8 @@ print("Chip Tower Examples")
 import example.combine
 import example.platformExample
 import example.ring
+import example.guard_rail
+import example.guard_rail_platform
 
 print('Walkway examples')
 import example.walkway