+++ src/cqindustry/__init__.py
@@ -12,3 +12,5 @@ from .Barrier import (
     cut_magnets,
     caution_stripe
 )
+from .GuardRail import GuardRail
+from .make_platform_rails import make_platform_rails